Extract links to a separate component
