+++ src/components/NavigationLink.vue
@@ -1,0 +1,26 @@
+<template>
+  <router-link
+    :to="`/${to}`"
+    :class="[
+      currentRoute === to ? 'border-sky-500 text-slate-900 dark:bg-slate-900 dark:text-white' : 'border-transparent text-slate-500 hover:border-slate-300 hover:text-slate-700 dark:text-slate-300 dark:hover:bg-slate-700 dark:hover:text-white',
+      'inline-flex items-center border-b-2 px-1 pt-1 text-sm font-medium'
+    ]"
+  >
+    {{ name }}
+  </router-link>
+</template>
+
+
+<script setup>
+import { computed } from 'vue'
+import { useRoute } from 'vue-router'
+
+defineProps ({
+    name : {type: String, required: true},
+    to   : {type: String, required: true}
+})
+
+const currentRoute = computed(() => {
+    return useRoute().name
+})
+</script>
